Remove context viewState - unnecessary complexity. Create separate stores for each route, as it seems to be SvelteKit's preferred way of operation (works best with how its fetch wrapper works).
