+++ src/lib/view-state/agency-sector-impact.ts
