+++ src/lib/view-state/index.ts
@@ -1,4 +1,4 @@
-import { createAgencySectorImpactStore } from '$lib/view-state/stores/agency-sector-impact';
+import { createAgencySectorImpactStore } from '$lib/view-state/agency-sector-impact';
 
 export const createViewState = () => {
   const agencySectorImpact = createAgencySectorImpactStore();
